+++ CHANGELOG.md
@@ -5,6 +5,10 @@ documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Fixed
+- Parsing a pipeline with variables keyword
+
 ## [0.0.39] - 2020-06-02
 ### Added
 - Service documentation